Changed command bean to not use weird double field.
