+++ src/main/kotlin/pl/dwojciechowski/ui/PLMPluginNotification.kt
@@ -4,7 +4,6 @@ import com.intellij.notification.NotificationDisplayType
 import com.intellij.notification.NotificationGroup
 import com.intellij.notification.NotificationType
 import com.intellij.openapi.project.Project
-import pl.dwojciechowski.model.CommandBean
 import javax.swing.Icon
 
 internal object PLMPluginNotification {
@@ -27,10 +26,6 @@ internal object PLMPluginNotification {
 
     fun settingsSaved(project: Project) {
         notify(project, "Settings Saved", PluginIcons.CONFIRMATION)
-    }
-
-    fun notifyCommandBeanChange(project: Project, commandBean: CommandBean){
-
     }
 
     fun notify(project: Project, text: String, icon: Icon) {
